refactor(vhd-lib): split tests
