+++ packages/vhd-lib/src/_createFooterHeader.integ.spec.js
@@ -1,0 +1,10 @@
+/* eslint-env jest */
+import { createFooter } from './_createFooterHeader'
+
+test('createFooter() does not crash', () => {
+  createFooter(104448, Math.floor(Date.now() / 1000), {
+    cylinders: 3,
+    heads: 4,
+    sectorsPerTrack: 17,
+  })
+})
